Move the state-only tests into the isolation-state folder
The tests for parsing a command, picking runs by target, and formatting the
run list all check isolation-state.ts, but they sat in the folder mirroring
isolation.ts. The mutation gate only runs a source's own mirrored folder, so
those assertions never ran against isolation-state mutants. Moving them takes
that file from 65.5% to 78.3% without a single new test. The clean-command
tests stay where they are, since they go through isolation.ts.

Also records the remaining clean race in TODO.md: a clean that decides a run is
abandoned and only then removes it can still delete a run that took the lock in
between. No record write can close that window — the clean has to hold the lock
across both steps.
